+++ backend/app/db/models.py
@@ -96,7 +96,7 @@ class Instrument(BaseDBModel):
     atr_profit_multiple_percentage: Mapped[Decimal] = mapped_column(
         nullable=False, default=Decimal("1.0")
     )
-    max_position_size: Mapped[int] = mapped_column(nullable=True)
+    max_position_size: Mapped[Decimal] = mapped_column(nullable=True)
     opening_price_multiple_percentage: Mapped[Decimal] = mapped_column(
         nullable=False, default=Decimal("1.0")
     )
